# packages/contracts/openapi/occ-api.openapi.json
{
  "components": {
...
    "securitySchemes": {
      "serviceApiKey": { "in": "header", "name": "x-api-key", "type": "apiKey" },
      "sessionCookie": { "in": "cookie", "name": "openclaw_occ.session_token", "type": "apiKey" }
    }
  },
  "info": { "title": "Development OCC API", "version": "0.1.0" },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/sign-in/email": {
      "post": {
        "description": "Authenticates a local account and issues a user session cookie.",
        "operationId": "signInEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "email": { "maxLength": 320, "minLength": 3, "type": "string" },
                  "password": { "maxLength": 128, "minLength": 12, "type": "string" }
                },
                "required": ["email", "password"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "OK",
                  "properties": {
                    "data": {
                      "additionalProperties": false,
                      "properties": { "authenticated": { "const": true, "type": "boolean" } },
                      "required": ["authenticated"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["data", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Unauthorized",
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" }
                      },
                      "required": ["code", "message"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Unauthorized"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Service Unavailable",
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" }
                      },
                      "required": ["code", "message"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [],
        "summary": "Sign in with email and password",
        "tags": ["Authentication"]
      }
    },
    "/api/auth/sign-out": {
      "post": {
        "description": "Revokes the current user session cookie.",
        "operationId": "signOut",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "OK",
                  "properties": {
                    "data": { "additionalProperties": true, "type": "object" },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["data", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Unauthorized",
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" }
                      },
                      "required": ["code", "message"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Unauthorized"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Service Unavailable",
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" }
                      },
                      "required": ["code", "message"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [{ "sessionCookie": [] }],
        "summary": "Sign out of the current session",
        "tags": ["Authentication"]
      }
    },
...
  },
  "security": [{ "sessionCookie": [] }, { "serviceApiKey": [] }]
}